feat(techdocs): merge reader components
