+++ plugins/techdocs/package.json
@@ -45,7 +45,7 @@
     "@backstage/plugin-catalog": "^0.10.0",
     "@backstage/plugin-catalog-react": "^0.9.0",
     "@backstage/plugin-search": "^0.7.3",
-    "@backstage/plugin-techdocs-addons": "^0.0.0",
+    "@backstage/techdocs-addons": "^0.0.0",
     "@backstage/theme": "^0.2.15",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
@@ -54,7 +54,9 @@
     "dompurify": "^2.2.9",
     "event-source-polyfill": "^1.0.25",
     "git-url-parse": "^11.6.0",
+    "jss": "~10.8.2",
     "lodash": "^4.17.21",
+    "react-helmet": "6.1.0",
     "react-router": "6.0.0-beta.0",
     "react-router-dom": "6.0.0-beta.0",
     "react-text-truncate": "^0.18.0",
