+++ plugins/techdocs/src/reader/components/TechDocsReaderPage/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2022 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,10 @@
  * limitations under the License.
  */
 
+export { TechDocsReaderPage, TechDocsReaderLayout } from './TechDocsReaderPage';
 export type {
   TechDocsReaderPageProps,
   TechDocsReaderLayoutProps,
 } from './TechDocsReaderPage';
-export {
-  TechDocsReaderLayout,
-  useTechDocsMetadata,
-  useEntityMetadata,
-  useTechDocsReaderPage,
-  useShadowRoot,
-  useShadowRootElements,
-} from './TechDocsReaderPage';
-export * from './TechDocsReaderPageHeader';
-export * from './TechDocsReaderPageContent';
-export * from './TechDocsStateIndicator';
+export * from './context';
+export * from './hooks';
