+++ plugins/techdocs/src/reader/components/TechDocsReaderPageContent/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2022 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,5 @@
  * limitations under the License.
  */
 
-export type {
-  TechDocsReaderPageProps,
-  TechDocsReaderLayoutProps,
-} from './TechDocsReaderPage';
-export {
-  TechDocsReaderLayout,
-  useTechDocsMetadata,
-  useEntityMetadata,
-  useTechDocsReaderPage,
-  useShadowRoot,
-  useShadowRootElements,
-} from './TechDocsReaderPage';
-export * from './TechDocsReaderPageHeader';
-export * from './TechDocsReaderPageContent';
-export * from './TechDocsStateIndicator';
+export { TechDocsReaderPageContent } from './TechDocsReaderPageContent';
+export * from './context';
